fix: AppHeader の story の作り方を修正
